Renamed into template/metaprogramming.h
