+++ template/docu/docu_05variadic_01.md
