+++ template/mapreduce.h
