+++ zzz_full_test/everything.cc
@@ -30,7 +30,7 @@
 #include "./rtti/docu/docu_06rtti_01_test.cc"
 #include "./rtti/test.cc"
 #include "./strings/test.cc"
+#include "./template/test.cc"
 #include "./time/test.cc"
 #include "./util/test.cc"
-#include "./variadic/test.cc"
 #include "./waitable_atomic/test.cc"
